factored out support for Alecto and KlikAanKlikUit devices
