+++ include/rtl_433_devices.h
@@ -14,5 +14,7 @@ extern r_device acurite_rain_gauge;
 extern r_device oregon_scientific;
 extern r_device mebus433;
 extern r_device intertechno;
+extern r_device newkaku;
+extern r_device alectov1;
 
 #endif /* INCLUDE_RTL_433_DEVICES_H_ */
